pyproject.toml
[project]
name = "email-assistant-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.187.0",
    "google-auth>=2.47.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.2",
    "grandalf>=0.8",
    "ipykernel>=7.1.0",
    "langchain>=1.2.3",
    "langchain-groq>=1.1.1",
    "langgraph>=1.0.5",
    "pandas>=2.3.3",
    "plotly>=6.5.2",
    "python-dotenv>=1.2.1",
    "streamlit>=1.53.0",
    "tavily-python>=0.7.19",
]
